pyproject.toml
[tool.poetry]
name = "soar-asset-mocker"
version = "0.1.0"
description = "Asset Mocker enhances SOAR Connectors with ability to record and replay all external commnication, allowing for faster playbook development"
authors = ["Michał Posłuszny <mposluszny@splunk.com>"]
readme = "README.md"
packages = [{include = "soar_asset_mocker"}]

[tool.poetry.dependencies]
python = "^3.9"
vcrpy = "6.0.1"
pyyaml = "6.0.1"
dacite = "^1.8.1"

[tool.poetry.group.dev.dependencies]
poetry = "^1.8.2"
pytest = "^8.3.2"
typer = "^0.9.0"
coverage = "^7.4.4"
pre-commit = { version = "^3.7.0", python = ">=3.6.1,<4.0" }
mypy = "^1.9.0"
types-PyYAML = "6.0.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[tool.black]
target-version = ['py39']
line-length = 79

[tool.isort]
profile = "black"
py_version=39
multi_line_output = 3
line_length = 79

[tool.coverage.report]
omit = [
    "tests/*",
    "__init__.py"
    ]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    ]
